Moved branch helpers into separate files under /utils/branch/
