+++ src/utils/branch/makeBranch.js
@@ -3,30 +3,31 @@ const assert = require('assert');
 
 const repoParentFolder = process.cwd() + '/../';
 
-module.exports = async function(api, repo, branchName, verbose = true) {
+module.exports = async function(repo, branchName, verbose = true) {
   console.log(`Making branch ${repo.name}/${branchName}.`);
   const repoFolder = `${repoParentFolder}${repo.name}`;
 
   // Get current branch and ensure that we're on master.
   const currentBranch = shelljs.exec(
     `
-    cd ${repoFolder} &&
-    git rev-parse --abbrev-ref HEAD
-  `,
+        cd ${repoFolder} &&
+        git rev-parse --abbrev-ref HEAD
+      `,
     {silent: !verbose}
   );
   const currBranchName = currentBranch.stdout.trim();
   console.log(' - Current branch name is:', currBranchName);
-  if (currBranchName !== 'master') {
-    throw new Error(`Branch for ${repo.name} is not master. Run setup.js.`);
-  }
+  assert(
+    currBranchName === 'master',
+    `Branch for ${repo.name} is not 'master'. Run setup.js.`
+  );
 
   // Create branch
   const branchCreate = shelljs.exec(
     `
-    cd ${repoFolder} &&
-    git checkout -b ${branchName}
-  `,
+        cd ${repoFolder} &&
+        git checkout -b ${branchName}
+      `,
     {silent: !verbose}
   );
   assert(
